% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{bike_get_mod_preds}
\alias{bike_get_mod_preds}
\title{Generate model predictions from a model}
\usage{
bike_get_mod_preds(mod, mod_name, test_df, pred_mat_func = NULL)
}
\arguments{
\item{mod}{Mode, fed to predict}

\item{mod_name}{model name for storage}

\item{test_df}{data frame of test data}

\item{pred_mat_func}{function to convert from test_df into prediction matrix for mod}
}
\value{
the test dataframe with predictions and residuals rbind-ed in
}
\description{
Generate model predictions from a model
}
