% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{bike_mod_results}
\alias{bike_mod_results}
\title{Get Bike Model Results}
\usage{
bike_mod_results(mod, mod_name, test_df, pred_mat_func)
}
\arguments{
\item{mod}{Mode, fed to predict}

\item{mod_name}{model name for storage}

\item{test_df}{data frame of test data}

\item{pred_mat_func}{function to convert from test_df into prediction matrix for mod}
}
\value{
None
}
\description{
Get Bike Model Results
}
