% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{bike_test_dat}
\alias{bike_test_dat}
\title{Get bike test data}
\usage{
bike_test_dat(con, split_date)
}
\arguments{
\item{con}{db connection}

\item{split_date}{date to split data (on or before is training, after is test)}
}
\value{
test bike data
}
\description{
Get bike test data
}
\examples{
con <- DBI::dbConnect(odbc::odbc(), "Content DB")
bike_test_dat(con, Sys.Date() - 2)
}
