% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{bike_train_dat}
\alias{bike_train_dat}
\title{Get Bike Training Data}
\usage{
bike_train_dat(con, split_date)
}
\arguments{
\item{con}{db connection}

\item{split_date}{date to split data (on or before is training, after is test)}
}
\value{
training dataset
}
\description{
Get Bike Training Data
}
\examples{
con <- DBI::dbConnect(odbc::odbc(), "Content DB")
bike_train_dat(con, "2019-12-05")
}
