% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{feeds_urls}
\alias{feeds_urls}
\title{Get dataframe of bikeshare feeds and URLs}
\usage{
feeds_urls(url = "https://gbfs.capitalbikeshare.com/gbfs/gbfs.json",
  lang = "en")
}
\arguments{
\item{url}{url of feeds json, has default}

\item{lang}{which language, defaults to english, "en"}
}
\value{
A dataframe of feed names and URLs
}
\description{
Get dataframe of bikeshare feeds and URLs
}
\examples{
if (interactive()) {
  feeds_urls()
}
}
