% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{prep_r_xgb_mat}
\alias{prep_r_xgb_mat}
\title{Turn test df into prediction matrix for R XGBoost model}
\usage{
prep_r_xgb_mat(df)
}
\arguments{
\item{df}{test data frame}
}
\value{
a matrix
}
\description{
Turn test df into prediction matrix for R XGBoost model
}
